# javascript: pooling and reuse with export functions + misc updates