--- a/pkg/js/libs/mysql/mysql.go
+++ b/pkg/js/libs/mysql/mysql.go
@@ -7,13 +7,12 @@ import (
 	"io"
 	"log"
 	"net"
-	"net/url"
 	"time"
 
 	"github.com/go-sql-driver/mysql"
 	"github.com/praetorian-inc/fingerprintx/pkg/plugins"
 	mysqlplugin "github.com/praetorian-inc/fingerprintx/pkg/plugins/services/mysql"
-	utils "github.com/projectdiscovery/nuclei/v3/pkg/js/utils"
+	"github.com/projectdiscovery/nuclei/v3/pkg/js/utils"
 	"github.com/projectdiscovery/nuclei/v3/pkg/protocols/common/protocolstate"
 )
 
@@ -21,16 +20,6 @@ import (
 //
 // Internally client uses go-sql-driver/mysql driver.
 type MySQLClient struct{}
-
-// Connect connects to MySQL database using given credentials.
-//
-// If connection is successful, it returns true.
-// If connection is unsuccessful, it returns false and error.
-//
-// The connection is closed after the function returns.
-func (c *MySQLClient) Connect(host string, port int, username, password string) (bool, error) {
-	return connect(host, port, username, password, "INFORMATION_SCHEMA")
-}
 
 // IsMySQL checks if the given host is running MySQL database.
 //
@@ -58,83 +47,96 @@ func (c *MySQLClient) IsMySQL(host string, port int) (bool, error) {
 	return true, nil
 }
 
-// ConnectWithDB connects to MySQL database using given credentials and database name.
+// Connect connects to MySQL database using given credentials.
 //
 // If connection is successful, it returns true.
 // If connection is unsuccessful, it returns false and error.
-//
 // The connection is closed after the function returns.
-func (c *MySQLClient) ConnectWithDB(host string, port int, username, password, dbName string) (bool, error) {
-	return connect(host, port, username, password, dbName)
+func (c *MySQLClient) Connect(host string, port int, username, password string) (bool, error) {
+	if !protocolstate.IsHostAllowed(host) {
+		// host is not valid according to network policy
+		return false, protocolstate.ErrHostDenied.Msgf(host)
+	}
+	dsn, err := BuildDSN(DSNOptions{
+		Host:     host,
+		Port:     port,
+		DbName:   "INFORMATION_SCHEMA",
+		Protocol: "tcp",
+		Username: username,
+		Password: password,
+	})
+	if err != nil {
+		return false, err
+	}
+	return connectWithDSN(dsn)
+}
+
+type MySQLInfo struct {
+	Host      string               `json:"host,omitempty"`
+	IP        string               `json:"ip"`
+	Port      int                  `json:"port"`
+	Protocol  string               `json:"protocol"`
+	TLS       bool                 `json:"tls"`
+	Transport string               `json:"transport"`
+	Version   string               `json:"version,omitempty"`
+	Debug     plugins.ServiceMySQL `json:"debug,omitempty"`
+	Raw       string               `json:"metadata"`
+}
+
+// returns MySQLInfo when fingerpint is successful
+func (c *MySQLClient) FingerprintMySQL(host string, port int) (MySQLInfo, error) {
+	info := MySQLInfo{}
+	if !protocolstate.IsHostAllowed(host) {
+		// host is not valid according to network policy
+		return info, protocolstate.ErrHostDenied.Msgf(host)
+	}
+	conn, err := protocolstate.Dialer.Dial(context.TODO(), "tcp", net.JoinHostPort(host, fmt.Sprintf("%d", port)))
+	if err != nil {
+		return info, err
+	}
+	defer conn.Close()
+
+	plugin := &mysqlplugin.MYSQLPlugin{}
+	service, err := plugin.Run(conn, 5*time.Second, plugins.Target{Host: host})
+	if err != nil {
+		return info, err
+	}
+	if service == nil {
+		return info, fmt.Errorf("something went wrong got null output")
+	}
+	// fill all fields
+	info.Host = service.Host
+	info.IP = service.IP
+	info.Port = service.Port
+	info.Protocol = service.Protocol
+	info.TLS = service.TLS
+	info.Transport = service.Transport
+	info.Version = service.Version
+	info.Debug = service.Metadata().(plugins.ServiceMySQL)
+	bin, _ := service.Raw.MarshalJSON()
+	info.Raw = string(bin)
+	return info, nil
 }
 
 // ConnectWithDSN connects to MySQL database using given DSN.
 // we override mysql dialer with fastdialer so it respects network policy
 func (c *MySQLClient) ConnectWithDSN(dsn string) (bool, error) {
+	return connectWithDSN(dsn)
+}
+
+func (c *MySQLClient) ExecuteQueryWithOpts(opts DSNOptions, query string) (*utils.SQLResult, error) {
+	if !protocolstate.IsHostAllowed(opts.Host) {
+		// host is not valid according to network policy
+		return nil, protocolstate.ErrHostDenied.Msgf(opts.Host)
+	}
+	dsn, err := BuildDSN(opts)
+	if err != nil {
+		return nil, err
+	}
+
 	db, err := sql.Open("mysql", dsn)
 	if err != nil {
-		return false, err
-	}
-	defer db.Close()
-	db.SetMaxOpenConns(1)
-	db.SetMaxIdleConns(0)
-
-	_, err = db.Exec("select 1")
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-func connect(host string, port int, username, password, dbName string) (bool, error) {
-	if host == "" || port <= 0 {
-		return false, fmt.Errorf("invalid host or port")
-	}
-
-	if !protocolstate.IsHostAllowed(host) {
-		// host is not valid according to network policy
-		return false, protocolstate.ErrHostDenied.Msgf(host)
-	}
-
-	target := net.JoinHostPort(host, fmt.Sprintf("%d", port))
-
-	db, err := sql.Open("mysql", fmt.Sprintf("%v:%v@tcp(%v)/%s?allowOldPasswords=1",
-		url.PathEscape(username),
-		url.PathEscape(password),
-		target,
-		dbName))
-	if err != nil {
-		return false, err
-	}
-	defer db.Close()
-	db.SetMaxOpenConns(1)
-	db.SetMaxIdleConns(0)
-
-	_, err = db.Exec("select 1")
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-// ExecuteQuery connects to Mysql database using given credentials and database name.
-// and executes a query on the db.
-func (c *MySQLClient) ExecuteQuery(host string, port int, username, password, dbName, query string) (string, error) {
-
-	if !protocolstate.IsHostAllowed(host) {
-		// host is not valid according to network policy
-		return "", protocolstate.ErrHostDenied.Msgf(host)
-	}
-
-	target := net.JoinHostPort(host, fmt.Sprintf("%d", port))
-
-	db, err := sql.Open("mysql", fmt.Sprintf("%v:%v@tcp(%v)/%s",
-		url.PathEscape(username),
-		url.PathEscape(password),
-		target,
-		dbName))
-	if err != nil {
-		return "", err
+		return nil, err
 	}
 	defer db.Close()
 	db.SetMaxOpenConns(1)
@@ -142,13 +144,31 @@ func (c *MySQLClient) ExecuteQuery(host string, port int, username, password, db
 
 	rows, err := db.Query(query)
 	if err != nil {
-		return "", err
+		return nil, err
 	}
-	resp, err := utils.UnmarshalSQLRows(rows)
+
+	data, err := utils.UnmarshalSQLRows(rows)
 	if err != nil {
-		return "", err
+		if len(data.Rows) > 0 {
+			// allow partial results
+			return data, nil
+		}
+		return nil, err
 	}
-	return string(resp), nil
+	return data, nil
+}
+
+// ExecuteQuery connects to Mysql database using given credentials and database name.
+// and executes a query on the db.
+func (c *MySQLClient) ExecuteQuery(host string, port int, username, password, dbName, query string) (*utils.SQLResult, error) {
+	return c.ExecuteQueryWithOpts(DSNOptions{
+		Host:     host,
+		Port:     port,
+		DbName:   dbName,
+		Protocol: "tcp",
+		Username: username,
+		Password: password,
+	}, query)
 }
 
 func init() {